+++ CHANGELOG.md
@@ -7,8 +7,8 @@
 ### First Release
 
 #### Module ryan_joiner
+- Added ``_critical_value`` function (06/11/2023);
 - Added ``p_value`` function (05/11/2023);
-- Added ``critical_value`` function (05/11/2023);
 - Added ``statistic`` function (05/11/2023);
 - Added ``citation`` function (04/11/2023);
 - Added ``normal_order_statistic`` function (02/11/2023);